bugfix
git-svn-id: svn://localhost:8080/bicluster/trunk@57 6e201009-ea7f-4747-b7f3-d4bff3160859

+++ src/org/caleydo/view/bicluster/sorting/ASortingStrategy.java
@@ -33,7 +33,7 @@ public abstract class ASortingStrategy implements Iterable<Integer> {
 
 
 
-	protected Set<Pair<Float, Integer>> indices;
+	protected Set<Pair<Integer, Float>> indices;
 
 	public abstract void setIndices(List<Integer> dimIndices);
